% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocFunctions.R
\name{Compute_AUC}
\alias{Compute_AUC}
\title{Computes the area under the curve of the specified type ROC}
\usage{
Compute_AUC(
  type = "TOC",
  model = "IO",
  mu = 1,
  sigmaT = 1,
  rho_0 = 0,
  rho_1 = 0,
  nSize = 4,
  method.Ens = NULL,
  lower = 0,
  upper = 1,
  alg = "simpson",
  maxIter = 100,
  tol = 0.001,
  ...
)
}
\arguments{
\item{type}{: String indicating the type of ROC for which the AUC will be computed. Default "TOC"}

\item{model}{: String indicating the model for which the Hit Rate will be computed. Default "IO"}

\item{mu}{: Mean of the target marginal distributions}

\item{sigmaT}{: Standard deviation of the target marginal distributions}

\item{rho_0}{: Value of the common correlation among non-target stimuli}

\item{rho_1}{: Value of the common correlation among non-target stimuli and the target}

\item{nSize}{: Number of simultaneously presented stimuli}

\item{method.Ens}{: Optional. Used to specify the method.Ens for the Ensemble model.}

\item{lower}{: Numeric. Lower value of pnorm(criterion) for which the area will be computed. Defaults to 0}

\item{upper}{: Numeric. Upper value of pnorm(criterion) for which the area will be computed. Defaults to 1}

\item{alg}{: String indicating the algorithm to approximate the AUC. Default "simpson"}

\item{maxIter}{: Numeric indicating maximum number of iteration of the algorithm at each recursion level.}

\item{tol}{: Numeric indicating maximum tolerated difference between iterations of the algorithm}

\item{...}{: Additional parameters to be passed to the specific model function}
}
\value{
numeric value of the AUC with tol attribute
}
\description{
Computes the area under the curve of the specified type ROC
}
\examples{
Compute_AUC(mu = 1, sigmaT = 1, nSize = 2)
          Compute_AUC(mu = 1, sigmaT = 1, nSize = 2)

}
