% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocFunctions.R
\name{Compute_Rates}
\alias{Compute_Rates}
\title{Computes the HR, FAR, RR_TA, and RR_TP of the model for a number of criterion values.}
\usage{
Compute_Rates(
  model = "IO",
  nSize = 4,
  mu = 1,
  sigmaT = 1,
  rho_0 = 0,
  rho_1 = 0,
  criterion = NULL,
  breaks = NULL,
  nSteps = 41,
  method.Ens = NULL,
  ...
)
}
\arguments{
\item{model}{: String indicating the model for which the Hit Rate will be computed. Default "IO"}

\item{nSize}{: Number of simultaneously presented stimuli}

\item{mu}{: Mean of the target marginal distributions}

\item{sigmaT}{: Standard deviation of the target marginal distributions}

\item{rho_0}{: Value of the common correlation among non-target stimuli}

\item{rho_1}{: Value of the common correlation among non-target stimuli and the target}

\item{criterion}{: Optional. Values of the criterion at which the rates will be computed.
If missing, breaks or nSteps will be used to determine the criteria.}

\item{breaks}{: Optional. Numeric values between 0 and 1. criterion values are created as qnorm(breaks)}

\item{nSteps}{: Used if neither criterion nor breaks is provided. criterion values are created as a vector
of length nSteps}

\item{method.Ens}{: Optional. Used to specify the method.Ens for the Ensemble model.}

\item{...}{: Additional parameters to be passed to the specific model function}
}
\value{
data.frame with columns criterion, HR, FAR, RRTP, and RRTA
}
\description{
Computes the HR, FAR, RR_TA, and RR_TP of the model for a number of criterion values.
}
\examples{
Compute_Rates(model = "IO",  mu = 1, sigmaT = 1, nSize = 2)
          Compute_Rates(model = "DO",  mu = 1, sigmaT = 1, nSize = 2)
          Compute_Rates(model = "Int", mu = 1, sigmaT = 1, nSize = 2)
          Compute_Rates(model = "Ens", mu = 1, sigmaT = 1, nSize = 2)

}
