% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomModel.R
\name{Custom_FAR}
\alias{Custom_FAR}
\title{Computes the probability of choosing the one non-target stimulus when no target is shown (False Alarm Rate) for
 Custom Models}
\usage{
Custom_FAR(criterion = 0.25, rho_0 = 0, nSize = 4, cMatrix, ...)
}
\arguments{
\item{criterion}{: Value of the criterion for selection}

\item{rho_0}{: Value of the common correlation among non-target stimuli}

\item{nSize}{: Number of simultaneously presented stimuli. Must equal ncol(cMatrix).}

\item{cMatrix}{: Matrix that defines the custom model}

\item{...}{: Captures unused parameter when called from generic function unforcedChoice_FAR()}
}
\value{
numeric value of the False Alarm Rate
}
\description{
Computes the probability of choosing the one non-target stimulus when no target is shown (False Alarm Rate) for
 Custom Models
}
\examples{
unforcedChoice:::Custom_FAR(nSize = 2, criterion = .25,
                                      cMatrix = matrix(c(1, 0), nrow = 1))
          unforcedChoice:::Custom_FAR(nSize = 2, criterion = .50,
                                      cMatrix = matrix(c(1, 0), nrow = 1))
          unforcedChoice:::Custom_FAR(nSize = 2, criterion = -Inf,
                                      cMatrix = matrix(c(1, 0), nrow = 1))
}
