% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomModel.R
\name{Custom_RR}
\alias{Custom_RR}
\title{Computes the probability of rejecting the set when (Rejection Rate) for
 Custom Models}
\usage{
Custom_RR(
  criterion = 0.25,
  mu = 1,
  sigmaT = 1,
  rho_0 = 0,
  rho_1 = 0,
  nSize = 4,
  cMatrix,
  ...
)
}
\arguments{
\item{criterion}{: Value of the criterion for selection}

\item{mu}{: Mean of the target marginal distributions}

\item{sigmaT}{: Standard deviation of the target marginal distributions}

\item{rho_0}{: Value of the common correlation among non-target stimuli}

\item{rho_1}{: Value of the common correlation among non-target stimuli and the target}

\item{nSize}{: Number of simultaneously presented stimuli. Must equal ncol(cMatrix).}

\item{cMatrix}{: Matrix that defines the custom model}

\item{...}{: Captures unused parameter when called from generic function unforcedChoice_RRTA()}
}
\value{
numeric value of the Hit Rate
}
\description{
Computes the probability of rejecting the set when (Rejection Rate) for
 Custom Models
}
\examples{
unforcedChoice:::Custom_RR(mu = 1, sigmaT = 1, nSize = 2,
                                     criterion = .25, cMatrix = diag(2))
          unforcedChoice:::Custom_RR(mu = 1, sigmaT = 1, nSize = 2,
                                     criterion = .50, cMatrix = diag(2))
          unforcedChoice:::Custom_RR(mu = 1, sigmaT = 1, nSize = 2,
                                     criterion = -Inf, cMatrix = diag(2))

}
