% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOModel.R
\name{DO_FAR}
\alias{DO_FAR}
\title{Computes the probability of choosing the one non-target stimulus when no target is shown (False Alarm Rate) for the
 Dependent Observations Model.}
\usage{
DO_FAR(criterion = 0.25, rho_0 = 0, nSize = 4, ...)
}
\arguments{
\item{criterion}{: Value of the criterion for selection}

\item{rho_0}{: Value of the common correlation among non-target stimuli}

\item{nSize}{: Number of simultaneously presented stimuli}

\item{...}{: Captures unused parameter when called from generic function unforcedChoice_FAR()}
}
\value{
numeric value of the False Alarm Rate
}
\description{
Computes the probability of choosing the one non-target stimulus when no target is shown (False Alarm Rate) for the
 Dependent Observations Model.
}
\examples{
unforcedChoice:::DO_FAR(nSize = 2, criterion = .25)
          unforcedChoice:::DO_FAR(nSize = 2, criterion = .50)
          unforcedChoice:::DO_FAR(nSize = 2, criterion = -Inf)
}
