% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOModel.R
\name{DO_RRTA}
\alias{DO_RRTA}
\title{Computes the probability of rejecting the set when no target is shown (Rejection Rate) for the
 Dependent Observations Model.}
\usage{
DO_RRTA(criterion = 0.25, rho_0 = 0, nSize = 4, ...)
}
\arguments{
\item{criterion}{: Value of the criterion for selection}

\item{rho_0}{: Value of the common correlation among non-target stimuli}

\item{nSize}{: Number of simultaneously presented stimuli}

\item{...}{: Captures unused parameter when called from generic function unforcedChoice_RRTA()}
}
\value{
numeric value of the Hit Rate
}
\description{
Computes the probability of rejecting the set when no target is shown (Rejection Rate) for the
 Dependent Observations Model.
}
\examples{
unforcedChoice:::DO_RRTP(nSize = 2, criterion = .25)
          unforcedChoice:::DO_RRTP(nSize = 2, criterion = .50)
          unforcedChoice:::DO_RRTP(nSize = 2, criterion = -Inf)

}
