% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsModel.R
\name{Ens_HR}
\alias{Ens_HR}
\title{Computes the probability of choosing the target stimulus when a target is shown (Hit Rate) for the
 Ensemble Model}
\usage{
Ens_HR(
  criterion = 0.25,
  mu = 1,
  sigmaT = 1,
  rho_0 = 0,
  rho_1 = 0,
  nSize = 4,
  method.Ens = "base",
  ...
)
}
\arguments{
\item{criterion}{: Value of the criterion for selection}

\item{mu}{: Mean of the target marginal distributions}

\item{sigmaT}{: Standard deviation of the target marginal distributions}

\item{rho_0}{: Value of the common correlation among non-target stimuli}

\item{rho_1}{: Value of the common correlation among non-target stimuli and the target}

\item{nSize}{: Number of simultaneously presented stimuli}

\item{method.Ens}{: Whether to integrate using base-R integrate (if set to "base" default. More precise, slower) or
mvtnorm::pmvnorm (if set to "mvnorm". Less precise, faster)}

\item{...}{: Captures unused parameter when called from generic function unforcedChoice_HR()}
}
\value{
numeric value of the Hit Rate
}
\description{
Computes the probability of choosing the target stimulus when a target is shown (Hit Rate) for the
 Ensemble Model
}
\examples{
unforcedChoice:::Ens_HR(mu = 1, sigmaT = 1, nSize = 2, criterion = .25)
          unforcedChoice:::Ens_HR(mu = 1, sigmaT = 1, nSize = 2, criterion = .50)
          unforcedChoice:::Ens_HR(mu = 1, sigmaT = 1, nSize = 2, criterion = -Inf)

}
