% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsModel.R
\name{Ens_RRTA}
\alias{Ens_RRTA}
\title{Computes the probability of rejecting the set when no target is shown (Rejection Rate) for the
 Ensemble Model}
\usage{
Ens_RRTA(criterion = 0.25, rho_0 = 0, nSize = 4, method.Ens = "base", ...)
}
\arguments{
\item{criterion}{: Value of the criterion for selection}

\item{rho_0}{: Value of the common correlation among non-target stimuli}

\item{nSize}{: Number of simultaneously presented stimuli}

\item{method.Ens}{: Whether to integrate using base-R integrate (if set to "base" default. More precise, slower) or
mvtnorm::pmvnorm (if set to "mvnorm". Less precise, faster)}

\item{...}{: Captures unused parameter when called from generic function unforcedChoice_RRTA()}
}
\value{
numeric value of the Hit Rate
}
\description{
Computes the probability of rejecting the set when no target is shown (Rejection Rate) for the
 Ensemble Model
}
\examples{
unforcedChoice:::Ens_RRTA(nSize = 2, criterion = .25)
          unforcedChoice:::Ens_RRTA(nSize = 2, criterion = .50)
          unforcedChoice:::Ens_RRTA(nSize = 2, criterion = -Inf)

}
