% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembleCriterionFunctions.R
\name{ID_ensembleCrit}
\alias{ID_ensembleCrit}
\title{ID_ensembleCrit
 Computes the hit and false alarm rates for the lineup
 ensemble criterion model.
 The rate is computed for the False Alarm Rate or the Hit Rate by setting
 the parameter mu equal to mu_0 or mu_1, the respective means of the
 Target Absent and Target Present distributions.}
\usage{
ID_ensembleCrit(
  criterion = 0.25,
  mu = 0,
  mu_0 = 0,
  sigma = 1,
  rho = 0,
  nSize = 4,
  method = "base"
)
}
\arguments{
\item{criterion}{: Value of the criterion for selection}

\item{mu}{: Mean of the marginal distribution}

\item{mu_0}{: Mean of the noise distributions}

\item{sigma}{: The common Standard deviation of the marginal distributions (Heteroscedasticity is not yet implemented)}

\item{rho}{: Value of the common correlation among stimuli}

\item{nSize}{: Number of simultaneously presented stimuli}

\item{method}{: Whether to integrate using base-R integrate (if set to "base" default. More precise, slower) or
mvtnorm::pmvnorm (if set to "mvnorm". Less precise, faster). Argument 'truncated' can also be used
to coompute Wixted et al. (2018) approximation (faster than 'base' but quite biased for small nSize).}
}
\value{
numeric value of the probability
}
\description{
ID_ensembleCrit
 Computes the hit and false alarm rates for the lineup
 ensemble criterion model.
 The rate is computed for the False Alarm Rate or the Hit Rate by setting
 the parameter mu equal to mu_0 or mu_1, the respective means of the
 Target Absent and Target Present distributions.
}
\examples{
ID_ensembleCrit(criterion = 1, mu = 0, sigma = 1, rho = 0, nSize = 2)
          ID_ensembleCrit(criterion = 1, mu = 1, sigma = 1, rho = 0, nSize = 2)

}
