% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntModel.R
\name{Int_FAR}
\alias{Int_FAR}
\title{Computes the probability of choosing the one non-target stimulus when no target is shown (False Alarm Rate) for the
 Integration Model.}
\usage{
Int_FAR(criterion = 0.25, rho_0 = 0, nSize = 4, ...)
}
\arguments{
\item{criterion}{: Value of the criterion for selection}

\item{rho_0}{: Value of the common correlation among non-target stimuli}

\item{nSize}{: Number of simultaneously presented stimuli}

\item{...}{: Captures unused parameter when called from generic function unforcedChoice_FAR()}
}
\value{
numeric value of the False Alarm Rate
}
\description{
Computes the probability of choosing the one non-target stimulus when no target is shown (False Alarm Rate) for the
 Integration Model.
}
\examples{
unforcedChoice:::Int_FAR(nSize = 2, criterion = .25)
          unforcedChoice:::Int_FAR(nSize = 2, criterion = .50)
          unforcedChoice:::Int_FAR(nSize = 2, criterion = -Inf)
}
