% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOModel.R
\name{IO_FAR}
\alias{IO_FAR}
\title{Computes the probability of choosing the one non-target stimulus when no target is shown (False Alarm Rate) for the
 Independent Observations Model.}
\usage{
IO_FAR(criterion = 0.25, nSize = 4, ...)
}
\arguments{
\item{criterion}{: Value of the criterion for selection}

\item{nSize}{: Number of simultaneously presented stimuli}

\item{...}{: Captures unused parameter when called from generic function unforcedChoice_FAR()}
}
\value{
numeric value of the False Alarm Rate
}
\description{
Computes the probability of choosing the one non-target stimulus when no target is shown (False Alarm Rate) for the
 Independent Observations Model.
}
\examples{
unforcedChoice:::IO_FAR(nSize = 2, criterion = .25)
          unforcedChoice:::IO_FAR(nSize = 2, criterion = .50)
          unforcedChoice:::IO_FAR(nSize = 2, criterion = -Inf)
}
