% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOModel.R
\name{IO_HR}
\alias{IO_HR}
\title{Computes the probability of choosing the target stimulus when a target is shown (Hit Rate) for the
 Independent Observations Model.}
\usage{
IO_HR(criterion = 0.25, mu = 1, sigmaT = 1, nSize = 4, ...)
}
\arguments{
\item{criterion}{: Value of the criterion for selection}

\item{mu}{: Mean of the target marginal distributions}

\item{sigmaT}{: Standard deviation of the target marginal distributions}

\item{nSize}{: Number of simultaneously presented stimuli}

\item{...}{: Captures unused parameter when called from generic function unforcedChoice_HR()}
}
\value{
numeric value of the Hit Rate
}
\description{
Computes the probability of choosing the target stimulus when a target is shown (Hit Rate) for the
 Independent Observations Model.
}
\examples{
unforcedChoice:::IO_HR(mu = 1, sigmaT = 1, nSize = 2, criterion = .25)
          unforcedChoice:::IO_HR(mu = 1, sigmaT = 1, nSize = 2, criterion = .50)
          unforcedChoice:::IO_HR(mu = 1, sigmaT = 1, nSize = 2, criterion = -Inf)

}
