% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOModel.R
\name{IO_RRTA}
\alias{IO_RRTA}
\title{Computes the probability of rejecting the set when no target is shown (Rejection Rate) for the
 Independent Observations Model.}
\usage{
IO_RRTA(criterion = 0.25, nSize = 4, ...)
}
\arguments{
\item{criterion}{: Value of the criterion for selection}

\item{nSize}{: Number of simultaneously presented stimuli}

\item{...}{: Captures unused parameter when called from generic function unforcedChoice_RRTA()}
}
\value{
numeric value of the Hit Rate
}
\description{
Computes the probability of rejecting the set when no target is shown (Rejection Rate) for the
 Independent Observations Model.
}
\examples{
unforcedChoice:::IO_RRTP(nSize = 2, criterion = .25)
          unforcedChoice:::IO_RRTP(nSize = 2, criterion = .50)
          unforcedChoice:::IO_RRTP(nSize = 2, criterion = -Inf)

}
