% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationRoutines.R
\name{parsInData}
\alias{parsInData}
\title{Plugs the values of pars in the data data.frame}
\usage{
parsInData(pars, fixed = NULL, data)
}
\arguments{
\item{pars}{: Numeric vector. Should contain the parameter values to assign to the unique values of
criterion, mu, sigmaT, rho_0, and rho_1, excluding those that appear on 'fixed'.}

\item{fixed}{: Optional character vector. Name of parameter that will fixed at values in data data.frame. Can include any of
'criterion', 'mu', 'sigmaT', 'rho_0', 'rho_1'}

\item{data}{: data.frame. Must have columns:
'model': character. Can be any of 'IO', 'DO', 'Ens', 'Int', 'Custom'
'nTP': numeric. Number of target present trials for the row.
'nTA': numeric. Number of target absent trials for the row.
'nSize': numeric. Number of stimuli presented on each trial for the row.
'criterion': If numeric and fixed includes 'criterion', values that will be used for computations for the row.
             Otherwise, it will be treated as a label of a criterion to be estimated equal across rows with the same value.
'mu': If numeric and fixed includes 'mu', values that will be used for computations for the row.
             Otherwise, it will be treated as a label of a mu to be estimated equal across rows with the same value.
'sigmaT': If numeric and fixed includes 'sigmaT', values that will be used for computations for the row.
             Otherwise, it will be treated as a label of a sigmaT to be estimated equal across rows with the same value.
'rho_0': If numeric and fixed includes 'rho_0', values that will be used for computations for the row.
             Otherwise, it will be treated as a label of a rho_0 to be estimated equal across rows with the same value.
'rho_1': If numeric and fixed includes 'rho_1', values that will be used for computations for the row.
             Otherwise, it will be treated as a label of a rho_1 to be estimated equal across rows with the same value.
'HR' : Numeric. Either a value between 0 and 1 given the Hit Rate or an integer giving the number of Hits for the row.
'FAR' : Numeric. Either a value between 0 and 1 given the False Alarm Rate or an integer giving the number of Misses for the row.
'RRTP' : Numeric. Either a value between 0 and 1 given the Rejection Rate on target present trials or an integer giving the number of Incorrect Rejections for the row.
'RRTA' : Numeric. Either a value between 0 and 1 given the Rejection Rate on target absent or an integer giving the number of Correct Rejections for the row.}
}
\value{
data.frame with pars replacing certain columns of data. See usage.
}
\description{
Plugs the values of pars in the data data.frame
}
