% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocFunctions.R
\name{plotROC}
\alias{plotROC}
\title{plotROC}
\usage{
plotROC(
  data = NULL,
  ratesTable = NULL,
  type = "ensemble",
  nSize = 4,
  mu_0 = 0,
  sigma_0 = 1,
  mu_1 = 1,
  sigma_1 = 1,
  rho = 0,
  title = NULL,
  subtitle = NULL,
  color = NULL,
  alpha = 0.3,
  nSteps = 41,
  method = NULL,
  fillerFAR = TRUE,
  AUC = TRUE,
  add = FALSE,
  ...
)
}
\arguments{
\item{data}{G}

\item{ratesTable}{G}

\item{type}{G}

\item{nSize}{G}

\item{mu_0}{G}

\item{sigma_0}{G}

\item{mu_1}{G}

\item{sigma_1}{G}

\item{rho}{G}

\item{title}{G}

\item{subtitle}{G}

\item{color}{G}

\item{alpha}{G}

\item{nSteps}{G}

\item{method}{G}

\item{fillerFAR}{G}

\item{AUC}{G}

\item{add}{G}

\item{...}{G G}
}
\value{
G
}
\description{
plotROC
}
\examples{
plotROC(AUC = FALSE)
}
