% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationRoutines.R
\name{predictData}
\alias{predictData}
\title{Predicts the HR, FAR, RRTP, and RRTA values the data data.frame (replaces those values is data if existing)}
\usage{
predictData(data, ...)
}
\arguments{
\item{data}{: data.frame. Must have columns:
'model': character. Can be any of 'IO', 'DO', 'Ens', 'Int', 'Custom'
'nTP': numeric. Number of target present trials for the row.
'nTA': numeric. Number of target absent trials for the row.
'nSize': numeric. Number of stimuli presented on each trial for the row.
'criterion': If numeric and fixed includes 'criterion', values that will be used for computations for the row.
             Otherwise, it will be treated as a label of a criterion to be estimated equal across rows with the same value.
'mu': If numeric and fixed includes 'mu', values that will be used for computations for the row.
             Otherwise, it will be treated as a label of a mu to be estimated equal across rows with the same value.
'sigmaT': If numeric and fixed includes 'sigmaT', values that will be used for computations for the row.
             Otherwise, it will be treated as a label of a sigmaT to be estimated equal across rows with the same value.
'rho_0': If numeric and fixed includes 'rho_0', values that will be used for computations for the row.
             Otherwise, it will be treated as a label of a rho_0 to be estimated equal across rows with the same value.
'rho_1': If numeric and fixed includes 'rho_1', values that will be used for computations for the row.
             Otherwise, it will be treated as a label of a rho_1 to be estimated equal across rows with the same value.}

\item{...}{: Additional arguments passed to Compute_Rates}
}
\value{
data.frame with pars replacing certain columns of data. See usage.
}
\description{
Predicts the HR, FAR, RRTP, and RRTA values the data data.frame (replaces those values is data if existing)
}
