% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembleCriterionFunctions.R
\name{rejection_ensembleCrit}
\alias{rejection_ensembleCrit}
\title{rejection_ensembleCrit
 Computes the probability of rejecting the lineup
 ensemble criterion model.
 The rate is computed for the Target Absent or the Target Present case by setting
 the parameter mu equal to mu_0 or mu_1, the respective means of the
 Target Absent and Target Present distributions.}
\usage{
rejection_ensembleCrit(
  criterion = 0.25,
  mu = 0,
  sigma = 1,
  rho = 0,
  nSize = 4,
  method = "base"
)
}
\arguments{
\item{criterion}{: Value of the criterion for selection}

\item{mu}{: Mean of the marginal distributions}

\item{sigma}{: The common Standard deviation of the marginal distributions (Heteroscedasticity is not yet implemented)}

\item{rho}{: Value of the common correlation among stimuli}

\item{nSize}{: Number of simultaneously presented stimuli}

\item{method}{: Whether to integrate using base-R integrate (if set to "base" default. More precise, slower) or
mvtnorm::pmvnorm (if set to "mvnorm". Less precise, faster)}
}
\value{
numeric value of the probability
}
\description{
rejection_ensembleCrit
 Computes the probability of rejecting the lineup
 ensemble criterion model.
 The rate is computed for the Target Absent or the Target Present case by setting
 the parameter mu equal to mu_0 or mu_1, the respective means of the
 Target Absent and Target Present distributions.
}
\examples{
rejection_ensembleCrit(criterion = 1, mu = 0, sigma = 1, rho = 0, nSize = 2)
          rejection_ensembleCrit(criterion = 1, mu = 1, sigma = 1, rho = 0, nSize = 2)

}
