% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembleCriterionFunctions.R
\name{rejection_TA_ensembleCrit}
\alias{rejection_TA_ensembleCrit}
\title{rejection_TA_ensembleCrit
 Computes the probability of rejecting the lineup when the target is absent for the
 ensemble criterion model.}
\usage{
rejection_TA_ensembleCrit(
  criterion = 0.25,
  mu_0 = 0,
  sigma_0 = 1,
  rho = 0,
  nSize = 4,
  method = "base"
)
}
\arguments{
\item{criterion}{: Value of the criterion for selection}

\item{mu_0}{: Mean of the marginal distributions of the non target stimuli}

\item{sigma_0}{: The common Standard deviation of the marginal distributions (Heteroscedasticity is not yet implemented)}

\item{rho}{: Value of the common correlation among stimuli}

\item{nSize}{: Number of simultaneously presented stimuli}

\item{method}{: Whether to integrate using base-R integrate (if set to "base" default. More precise, slower) or
mvtnorm::pmvnorm (if set to "mvnorm". Less precise, faster)}
}
\value{
numeric value of the probability
}
\description{
rejection_TA_ensembleCrit
 Computes the probability of rejecting the lineup when the target is absent for the
 ensemble criterion model.
}
\examples{
rejection_TA_ensembleCrit(criterion = 1, mu_0 = 0, sigma_0 = 1, rho = 0, nSize = 2)

}
