% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionCriterionFunctions.R
\name{rejection_TP_projectionCrit}
\alias{rejection_TP_projectionCrit}
\title{rejection_TP_projectionCrit
 Computes the probability of rejecting the lineup when the target is present for the
 projection criterion models (Independent Observations, Wixted, and Dependent Observations, Melisa)}
\usage{
rejection_TP_projectionCrit(
  criterion = 0.25,
  mu_0 = 0,
  mu_1 = 1,
  sigma_0 = 1,
  sigma_1 = 1,
  rho = 0,
  nSize = 4,
  method = "cuhre",
  ...
)
}
\arguments{
\item{criterion}{: Value of the criterion for selection}

\item{mu_0}{: Mean of the target absent marginal distributions}

\item{mu_1}{: Mean of the target present marginal distributions}

\item{sigma_0}{: Standard deviation of the target absent marginal distributions}

\item{sigma_1}{: Standard deviation of the target present marginal distributions}

\item{rho}{: Value of the common correlation among stimuli latent}

\item{nSize}{: Number of simultaneously presented stimuli}

\item{method}{: Integration method to pass to cubature::cubintegrate}

\item{...}{: Additional (optional) parameters to pass to cubature::cubintegrate}
}
\value{
numeric value of the probability or list from cubature::cubintegrate if integration fails (non-zero returnCode)
}
\description{
rejection_TP_projectionCrit
 Computes the probability of rejecting the lineup when the target is present for the
 projection criterion models (Independent Observations, Wixted, and Dependent Observations, Melisa)
}
\examples{
rejection_TP_projectionCrit(mu_0 = 0, sigma_0 = 1, rho = 0, nSize = 2, criterion = .25)
          rejection_TP_projectionCrit(mu_0 = 0, sigma_0 = 1, rho = 0, nSize = 2, criterion = .50)
          rejection_TP_projectionCrit(mu_0 = 0, sigma_0 = 1, rho = 0, nSize = 2, criterion = -Inf)

}
