% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionCriterionFunctions.R
\name{target_ID_TP_projectionCrit}
\alias{target_ID_TP_projectionCrit}
\title{target_ID_TP_projectionCrit
 Computes the probability of choosing the target stimulus when target is shown in the lineup for the
 projection criterion models (Independent Observations, Wixted, and Dependent Observations, Melisa)
 This provides the Hit Rate.}
\usage{
target_ID_TP_projectionCrit(
  criterion = 0.25,
  mu_0 = 0,
  mu_1 = 1,
  sigma_0 = 1,
  sigma_1 = 1,
  rho = 0,
  nSize = 4,
  method = "cuhre",
  ...
)
}
\arguments{
\item{criterion}{: Value of the criterion for selection}

\item{mu_0}{: Mean of the target absent marginal distributions}

\item{mu_1}{: Mean of the target present marginal distributions}

\item{sigma_0}{: Standard deviation of the target absent marginal distributions}

\item{sigma_1}{: Standard deviation of the target present marginal distributions}

\item{rho}{: Value of the common correlation among stimuli latent}

\item{nSize}{: Number of simultaneously presented stimuli}

\item{method}{: Integration method to pass to cubature::cubintegrate}

\item{...}{: Additional (optional) parameters to pass to cubature::cubintegrate}
}
\value{
numeric value of the probability or list from cubature::cubintegrate if integration fails (non-zero returnCode)
}
\description{
target_ID_TP_projectionCrit
 Computes the probability of choosing the target stimulus when target is shown in the lineup for the
 projection criterion models (Independent Observations, Wixted, and Dependent Observations, Melisa)
 This provides the Hit Rate.
}
\examples{
target_ID_TP_projectionCrit(mu_0 = 0, sigma_0 = 1, rho = 0, nSize = 2, criterion = .25)
          target_ID_TP_projectionCrit(mu_0 = 0, sigma_0 = 1, rho = 0, nSize = 2, criterion = .50)
          target_ID_TP_projectionCrit(mu_0 = 0, sigma_0 = 1, rho = 0, nSize = 2, criterion = -Inf)

}
