% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toyInt}
\alias{toyInt}
\title{Example data that perfectly follows the Integration model}
\format{
data.frame
}
\usage{
data(toyInt)
}
\description{
data.frame with rates of an Integration model
}
