% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unforcedChoice.R
\name{unforcedChoice_HR}
\alias{unforcedChoice_HR}
\title{Computes the probability of choosing the target stimulus when a target is shown (Hit Rate) for the
 model in the argument. Default is Independent Observations Model (model = "IO")}
\usage{
unforcedChoice_HR(
  criterion = 0.25,
  mu = 1,
  sigmaT = 1,
  rho_0 = 0,
  rho_1 = 0,
  nSize = 4,
  model = "IO",
  ...
)
}
\arguments{
\item{criterion}{: Value of the criterion for selection}

\item{mu}{: Mean of the target marginal distributions}

\item{sigmaT}{: Standard deviation of the target marginal distributions}

\item{rho_0}{: Value of the common correlation among non-target stimuli}

\item{rho_1}{: Value of the common correlation among non-target stimuli and the target}

\item{nSize}{: Number of simultaneously presented stimuli}

\item{model}{: String indicating the model for which the Hit Rate will be computed. Default "IO"}

\item{...}{: Additional parameters to be passed to the specific model function}
}
\value{
numeric value of the Hit Rate
}
\description{
Computes the probability of choosing the target stimulus when a target is shown (Hit Rate) for the
 model in the argument. Default is Independent Observations Model (model = "IO")
}
\examples{
unforcedChoice_HR(mu = 1, sigmaT = 1, nSize = 2, criterion = .25)
          unforcedChoice_HR(mu = 1, sigmaT = 1, nSize = 2, criterion = .50)
          unforcedChoice_HR(mu = 1, sigmaT = 1, nSize = 2, criterion = -Inf)

}
